remove commented out code and make update
